+++ sunpy/database/tests/test_database.py
@@ -22,7 +22,7 @@ from sunpy.database.tables import DatabaseEntry, Tag, FitsHeaderEntry,\
 from sunpy.database.commands import EmptyCommandStackError, NoSuchEntryError
 from sunpy.database.caching import LRUCache, LFUCache
 from sunpy.database import attrs
-from sunpy.net import vso
+from sunpy.net import vso, hek
 from sunpy.data.sample import RHESSI_EVENT_LIST
 from sunpy.data.test.waveunit import waveunitdir
 from sunpy.io import fits
@@ -48,6 +48,12 @@ def query_result():
     return vso.VSOClient().query(
         vso.attrs.Time('20130801T200000', '20130801T200030'),
         vso.attrs.Instrument('PLASTIC'))
+
+@pytest.fixture
+def download_qr():
+    return vso.VSOClient().query(
+        vso.attrs.Time('2012-03-29', '2012-03-29'),
+        vso.attrs.Instrument('AIA'))
 
 
 @pytest.fixture
@@ -353,6 +359,8 @@ def test_add_many(database):
     database.add_many((DatabaseEntry() for _ in xrange(5)))
     assert len(database) == 5
     database.undo()
+    with pytest.raises(EmptyCommandStackError):
+        database.undo()
     assert len(database) == 0
     database.redo()
     assert len(database) == 5
@@ -388,6 +396,36 @@ def test_add_already_existing_entry_ignore(database):
     database.add(entry, True)
     database.commit()
     assert entry.id == 1
+
+
+@pytest.mark.online
+def test_add_entry_from_hek_qr(database):
+    hek_res = hek.HEKClient().query(
+        hek.attrs.Time('2011/08/09 07:23:56', '2011/08/09 07:24:00'),
+        hek.attrs.EventType('FL'))
+    assert len(database) == 0
+    database.add_from_hek_query_result(hek_res)
+    assert len(database) == 2133
+
+
+@pytest.mark.online
+@pytest.mark.skipif(
+        sys.version_info[:2] == (2,6),
+        reason='for some unknown reason, this test fails on Python 2.6')
+def test_download_from_qr(database, download_qr, tmpdir):
+    assert len(database) == 0
+    database.download_from_vso_query_result(
+        download_qr, path=str(tmpdir.join('{file}.fits')))
+    fits_pattern = str(tmpdir.join('*.fits'))
+    num_of_fits_headers = sum(
+        len(fits.get_header(file)) for file in glob.glob(fits_pattern))
+    assert len(database) == num_of_fits_headers > 0
+    for entry in database:
+        assert os.path.dirname(entry.path) == str(tmpdir)
+    database.undo()
+    assert len(database) == 0
+    database.redo()
+    assert len(database) == num_of_fits_headers > 0
 
 
 @pytest.mark.online
@@ -474,6 +512,22 @@ def test_edit_entry(database):
     assert entry.id == 1
     database.edit(entry, id=42)
     assert entry.id == 42
+
+
+def test_remove_many_entries(filled_database):
+    bar = Tag('bar')
+    bar.id = 2
+    # required to check if `remove_many` adds any entries to undo-history
+    filled_database.clear_histories()
+    filled_database.remove_many(filled_database[:8])
+    assert len(filled_database) == 2
+    assert list(filled_database) == [
+        DatabaseEntry(id=9),
+        DatabaseEntry(id=10, tags=[bar])]
+    filled_database.undo()
+    assert len(filled_database) == 10
+    with pytest.raises(EmptyCommandStackError):
+        filled_database.undo()
 
 
 def test_remove_existing_entry(database):