+++ sunpy/lightcurve/sources/lyra.py
@@ -4,6 +4,7 @@ from __future__ import absolute_import
 
 import datetime
 import urlparse
+import sys
 
 from matplotlib import pyplot as plt
 from astropy.io import fits
@@ -11,6 +12,7 @@ import pandas
 
 from sunpy.lightcurve import LightCurve 
 from sunpy.time import parse_time
+from sunpy.util.odict import OrderedDict
 
 __all__ = ['LYRALightCurve']
 
@@ -54,8 +56,9 @@ class LYRALightCurve(LightCurve):
 
         """Shows a plot of all four light curves"""
         figure = plt.figure()
+        plt.subplots_adjust(left=0.17,top=0.94,right=0.94,bottom=0.15)
         axes = plt.gca()
-
+        
         axes = self.data.plot(ax=axes, subplots=True, sharex=True, **kwargs)
         #plt.legend(loc='best')
 
@@ -63,11 +66,13 @@ class LYRALightCurve(LightCurve):
             if names < 3:
                 name = lyranames[names][i]
             else:
-                name = lyranames[0][i] + ' (' + lyranames[1][i] + ')'
-            axes[i].set_ylabel("%s (%s)" % (name, "W/m**2"))
+                name = lyranames[0][i] + ' \n (' + lyranames[1][i] + ')'
+            axes[i].set_ylabel( "%s %s" % (name, "\n (W/m**2)"),fontsize=9.5)
 
         axes[0].set_title("LYRA ("+ self.data.index[0].strftime('%Y-%m-%d') +")")
         axes[-1].set_xlabel("Time")
+        for axe in axes:
+            axe.locator_params(axis='y',nbins=6)
 
         figure.show()
 
@@ -120,9 +125,13 @@ class LYRALightCurve(LightCurve):
         table = {}
 
         for i, col in enumerate(fits_record.columns[1:-1]):
-            table[col.name] = fits_record.field(i + 1)
+            #temporary patch for big-endian data bug on pandas 0.13
+            if fits_record.field(i+1).dtype.byteorder == '>' and sys.byteorder =='little':
+                table[col.name] = fits_record.field(i + 1).byteswap().newbyteorder()
+            else:
+                table[col.name] = fits_record.field(i + 1)
 
         # Return the header and the data
-        return hdulist[0].header, pandas.DataFrame(table, index=times)
+        return OrderedDict(hdulist[0].header), pandas.DataFrame(table, index=times)