+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+Latest
+------
+  - changed _maps attribute in MapCube to a non-hidden type
+  - Added Nobeyama Radioheliograph data support to Lightcurve object.
+  - Fixed some tests on map method to support Windows
+  - Added a window/split method to time range
+  - Updates to spectrogram documentation
+  - Added method Database.add_from_hek_query_result to HEK database
+  - Added method Database.download_from_vso_query_result
+
 0.4.0
 -----
 Features:
@@ -18,6 +28,7 @@ Features:
 Bug Fixes:
     * Add support for VSO queries through proxies.
     * Fix apparent Right Ascension calulations.
+    * LightCurve meta data member now an OrderedDict Instance
 
 0.3.2
 -----